requirements.txt
python-chess 
tensorflow 
torch 
ray[rllib]
stockfish 
pandas
tqdm
stable-baselines3[extra]
gymnasium